Refine wrac_log integration
